Fixed delete friend bug
