+++ src/components/ProfileImgPicker.js
@@ -1,22 +1,25 @@
 import React, { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
-import * as ImagePicker from 'expo-image-picker';
 import {
     StyleSheet, 
     TouchableOpacity,
     View,
     Image,
 } from 'react-native';
-import { uploadImg, deleteProfileImg } from '../firebase/ProfileImgAPI';
+import { useSelector } from 'react-redux';
+import * as ImagePicker from 'expo-image-picker';
+import { 
+    uploadImg, 
+    deleteProfileImg 
+} from '../firebase/ProfileImgAPI';
 import { getCurrentUser } from '../firebase/loginAPI';
-import ProfileImgPickerModal from './ProfileImgPickerModal';
 import { MaterialIcons } from '@expo/vector-icons';
+import ProfileImgPickerModal from './ProfileImgPickerModal';
 
 const ProfileImgPicker = () => {
     const NO_PROFILE_IMG = "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png";
 
     const [photoURI, setPhotoURI] = useState(NO_PROFILE_IMG);
-    const [visible, setVisible] = useState(false);
+    const [modalVisible, setModalVisible] = useState(false);
     
     const profilePictures = useSelector(state => state.users.profilePictures);
     
@@ -38,7 +41,7 @@ const ProfileImgPicker = () => {
                 setPhotoURI(result.uri);
                 
                 await uploadImg(result.uri);
-                setVisible(false);
+                setModalVisible(false);
             }
         } catch (error) {
             console.log(error.message);
@@ -63,7 +66,7 @@ const ProfileImgPicker = () => {
                 setPhotoURI(result.uri);
                 
                 await uploadImg(result.uri);
-                setVisible(false);
+                setModalVisible(false);
             }
         } catch (error) {
             console.log(error.message);
@@ -74,14 +77,14 @@ const ProfileImgPicker = () => {
         await deleteProfileImg();
         
         setPhotoURI(NO_PROFILE_IMG);
-        setVisible(false);
+        setModalVisible(false);
     }
     
     return (
         <View>
             <TouchableOpacity
                 style={styles.photoFrame}
-                onPress={() => setVisible(true)}>
+                onPress={() => setModalVisible(true)}>
                 <Image
                     style={styles.profileImg} 
                     source={{ uri: photoURI }}
@@ -95,8 +98,8 @@ const ProfileImgPicker = () => {
                 </View>
             </TouchableOpacity>
             <ProfileImgPickerModal
-                isVisible={visible}
-                onClose={() => setVisible(false)}
+                isVisible={modalVisible}
+                onClose={() => setModalVisible(false)}
                 onImageLibraryPress={onImageLibraryPress}
                 onCameraPress={onCameraPress}
                 onDeletePress={onDeletePress}
